+++ src/flat_tree.rs
@@ -406,6 +406,13 @@ impl Tree {
             if !line.is_selectable() {
                 continue;
             }
+            // Not very elegant
+            // If we implement several types of search (path, content),
+            // we'll rather store a direct_match boolean in the line.
+            if self.options.pattern.is_some()
+                && self.options.pattern.score_of(&line.name).is_none() {
+                continue;
+            }
             if line.score > 0 {
                 self.selection = idx;
                 return true;