+++ src/cli.rs
@@ -26,6 +26,7 @@ pub struct AppLaunchArgs {
     pub install: bool,                   // installation is required
     pub height: Option<u16>,             // an optional height to replace the screen's one
     pub no_style: bool,                  // whether to remove all styles (including colors)
+    pub specific_conf: Option<Vec<PathBuf>>,
 }
 
 #[cfg(not(windows))]
@@ -93,6 +94,8 @@ pub fn read_launch_args() -> Result<AppLaunchArgs, ProgramError> {
         tree_options.show_hidden = true;
         tree_options.respect_git_ignore = false;
     }
+    let specific_conf = cli_args.value_of("conf")
+        .map(|s| s.split(';').map(PathBuf::from).collect());
     Ok(AppLaunchArgs {
         root,
         file_export_path,
@@ -104,6 +107,7 @@ pub fn read_launch_args() -> Result<AppLaunchArgs, ProgramError> {
         install,
         height,
         no_style,
+        specific_conf,
     })
 }