+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 
-### master
-- fix panic on input being "///" (Fix #175)
+<a name="v0.12.1"></a>
+### v0.12.1 - 2020-01-25
+- fix panic on some inputs starting with a `/` (Fix #175)
 - TAB key now jumps to direct matches only
+- `--conf` arg to launch broot with specific config file(s) (fix #141)
 
 <a name="v0.12.0"></a>
 ### v0.12.0 - 2020-01-19